+++ src/pages/follow/Follow.jsx
@@ -16,7 +16,6 @@ import {
 	Wrapper,
 } from './follow.style';
 import { FollowButton } from '../../components/button/button.style';
-import userNoneProfile from '../../assets/image/profilePic.png';
 import FollowUnknown from './FollowUnknown';
 import Loading from '../../components/loading/Loading';
 import { Helmet } from 'react-helmet';
@@ -27,16 +26,17 @@ import {
 	useQueryClient,
 } from '@tanstack/react-query';
 import { delUnFollow, getFollow, postFollow } from '../../api/followApi';
+import FollowItem from './FollowItem';
 
 export default function Follwers() {
 	const navigate = useNavigate();
+	const queryClient = useQueryClient();
 	const accountUsername = useParams().accountUsername;
 	const followPage = useParams().follow;
 	const myAccountName = localStorage.getItem('userAccountName');
 	const count = useRef(0);
 	const [ref, inView] = useInView();
 	const [hasNextPage, setHasNextPage] = useState(true);
-	const queryClient = useQueryClient();
 
 	const {
 		data: followData,
@@ -71,37 +71,6 @@ export default function Follwers() {
 		queryClient.removeQueries({ queryKey: 'getFollowData' });
 	}, []);
 
-	const handleFollowChange = async (accountname, isfollow, e) => {
-		e.preventDefault();
-		if (isfollow) {
-			delFollowMutaion.mutate(accountname);
-		} else {
-			postFollowMutaion.mutate(accountname);
-		}
-	};
-
-	const delFollowMutaion = useMutation(delUnFollow, {
-		onSuccess: () => {
-			queryClient.invalidateQueries('profileData');
-		},
-		onError: () => {
-			console.error('실패');
-		},
-	});
-
-	const postFollowMutaion = useMutation(postFollow, {
-		onSuccess: () => {
-			queryClient.invalidateQueries('profileData');
-		},
-		onError: () => {
-			console.error('실패');
-		},
-	});
-
-	const handleImgError = (e) => {
-		e.target.src = userNoneProfile;
-	};
-
 	return (
 		<>
 			<Helmet>
@@ -123,52 +92,7 @@ export default function Follwers() {
 				{followData?.pages[0].data.length > 0
 					? followData?.pages.map((page) =>
 							page.data.map((follow) => {
-								return (
-									<UserWrap key={follow._id}>
-										<UserFlexWrap>
-											<UserProfileImg
-												onClick={() => {
-													myAccountName === follow.accountname
-														? navigate('../../../profile')
-														: navigate(`../../${follow.accountname}`);
-												}}
-											>
-												<UserFollowImage
-													src={follow.image}
-													onError={handleImgError}
-													alt={`${follow.username} 프로필 이미지입니다.`}
-												/>
-											</UserProfileImg>
-											<UserContent
-												onClick={() => {
-													myAccountName === follow.accountname
-														? navigate('../../../profile')
-														: navigate(`../../${follow.accountname}`);
-												}}
-											>
-												<UserFollowNickName>
-													{follow.username}
-												</UserFollowNickName>
-												<UserFollowIntro>{follow.intro}</UserFollowIntro>
-											</UserContent>
-										</UserFlexWrap>
-										{!(myAccountName === follow.accountname) && (
-											<FollowButton
-												type='button'
-												follow={follow.isfollow}
-												onClick={(e) => {
-													handleFollowChange(
-														follow.accountname,
-														follow.isfollow,
-														e
-													);
-												}}
-											>
-												{follow.isfollow === true ? '취소' : '팔로우'}
-											</FollowButton>
-										)}
-									</UserWrap>
-								);
+								return <FollowItem key={follow._id} follow={follow} />;
 							})
 					  )
 					: !isLoading && <FollowUnknown />}
